requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
fastapi==0.115.5
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
ratelim==0.1.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
